requirements.txt
# AI Workspace Requirements - CUDA Only
# Core ML Framework (install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121)
torch>=2.5.1+cu121
torchvision>=0.20.1+cu121
torchaudio>=2.5.1+cu121

# HTTP and API
requests>=2.31.0
aiohttp>=3.9.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Configuration Management
pyyaml>=6.0

# System Monitoring
psutil>=5.9.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0